+++ src/index.ts
@@ -4,7 +4,7 @@ export { configureMemiBoard, useMemiBoardConfig } from './config'
 export type { MemiBoardConfig, MemiBoardAuthOptions, MemiBoardModerationOptions } from './config'
 
 export { useMemiBoardAuth } from './composables/useMemiBoardAuth'
-export { useMemiBoardPosts } from './composables/useMemiBoardPosts'
+export { useMemiBoardPosts, useMemiBoardPostList, useMemiBoardPost } from './composables/useMemiBoardPosts'
 export type { CreatePostInput, UpdatePostInput } from './composables/useMemiBoardPosts'
 export { COMMENT_BODY_MAX_LENGTH, useMemiBoardComments, useMemiBoardReplies } from './composables/useMemiBoardComments'
 export type { AddCommentInput, AddReplyInput } from './composables/useMemiBoardComments'